+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Add
 
 - Video component on event slug, active and result event
+- og site name on event slug
 
 ## [0.0.81] - 2024-11-06